% tinyrox says don't edit this manually, but it can't stop you!
\name{default_media_key}
\alias{default_media_key}
\title{The default media reference key (\code{"DEFAULT_MEDIA"})}
\usage{
default_media_key()
}
\value{
The string \code{"DEFAULT_MEDIA"}.
}
\description{
The default media reference key (\code{"DEFAULT_MEDIA"})
}
\examples{
default_media_key()
}
