% tinyrox says don't edit this manually, but it can't stop you!
\name{from_json_file}
\alias{from_json_file}
\title{Read an OTIO JSON file into the object model}
\usage{
from_json_file(file_name)
}
\arguments{
\item{file_name}{Path to a \code{.otio} JSON file.}
}
\value{
The reconstructed OTIO object.
}
\description{
Read an OTIO JSON file into the object model
}
