% tinyrox says don't edit this manually, but it can't stop you!
\name{from_seconds}
\alias{from_seconds}
\title{Construct a RationalTime from seconds at a rate}
\usage{
from_seconds(seconds, rate = 1)
}
\arguments{
\item{seconds}{Numeric seconds.}

\item{rate}{Rate (fps).}
}
\description{
Construct a RationalTime from seconds at a rate
}
