% tinyrox says don't edit this manually, but it can't stop you!
\name{from_topleft}
\alias{from_topleft}
\title{Convert canonical top-left (pos_x, pos_y) to a user coord system}
\usage{
from_topleft(pos_x, pos_y, coord, canvas_w, canvas_h, displayed_w, displayed_h)
}
\arguments{
\item{coord}{User coord system: \code{"topleft"}, \code{"cartesian"},
or \code{"center"}.}
}
\description{
Inverse of \code{\link{to_topleft}}.
}
