% tinyrox says don't edit this manually, but it can't stop you!
\name{has_child}
\alias{has_child}
\title{Does a composition directly contain a child?}
\usage{
has_child(x, child)
}
\arguments{
\item{x}{A composition or collection.}

\item{child}{An OTIO object.}
}
\description{
Does a composition directly contain a child?
}
