% tinyrox says don't edit this manually, but it can't stop you!
\name{index_of_child}
\alias{index_of_child}
\title{1-based position of a child, or NA}
\usage{
index_of_child(x, child)
}
\arguments{
\item{x}{A composition or collection.}

\item{child}{An OTIO object.}
}
\description{
1-based position of a child, or NA
}
