% tinyrox says don't edit this manually, but it can't stop you!
\name{is_missing_reference}
\alias{is_missing_reference}
\title{Is x a MissingReference?}
\usage{
is_missing_reference(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
\code{TRUE} if \code{x} is a \code{MissingReference}, else \code{FALSE}.
}
\description{
Is x a MissingReference?
}
\examples{
is_missing_reference(MissingReference())
}
