% tinyrox says don't edit this manually, but it can't stop you!
\name{is_otio}
\alias{is_otio}
\alias{is_timeline}
\alias{is_composition}
\alias{is_media_reference}
\title{Type predicates for OTIO objects}
\usage{
is_otio(x)

is_timeline(x)

is_composition(x)

is_media_reference(x)
}
\arguments{
\item{x}{Object to test.}
}
\description{
Type predicates for OTIO objects
}
