% tinyrox says don't edit this manually, but it can't stop you!
\name{is_parent_of}
\alias{is_parent_of}
\title{Is x an ancestor of other?}
\usage{
is_parent_of(x, other)
}
\arguments{
\item{x}{A composition or collection.}

\item{other}{An OTIO object.}
}
\description{
Is x an ancestor of other?
}
