% tinyrox says don't edit this manually, but it can't stop you!
\name{kind}
\alias{kind}
\alias{kind<-}
\title{Get or set a track's kind}
\usage{
kind(x)

kind(x) <- value
}
\arguments{
\item{x}{A \code{\link{Track}}.}

\item{value}{\code{"Video"} or \code{"Audio"}.}
}
\description{
Get or set a track's kind
}
