% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{nle.api-package}
\alias{nle.api}
\alias{nle.api-package}
\title{\packageTitle{nle.api}}
\description{\packageDescription{nle.api}}
\section{Package Content}{\packageIndices{nle.api}}
\author{\packageAuthor{nle.api}}
\section{Maintainer}{\packageMaintainer{nle.api}}
\keyword{internal}
