% tinyrox says don't edit this manually, but it can't stop you!
\name{number_of_images_in_sequence}
\alias{number_of_images_in_sequence}
\title{Number of images in an ImageSequenceReference}
\usage{
number_of_images_in_sequence(x)
}
\arguments{
\item{x}{An \code{\link{ImageSequenceReference}}.}
}
\description{
Number of images in an ImageSequenceReference
}
