% tinyrox says don't edit this manually, but it can't stop you!
\name{otio_name}
\alias{otio_name}
\alias{otio_kind}
\title{Name / kind of an OTIO object}
\usage{
otio_name(x)

otio_kind(x)
}
\arguments{
\item{x}{An \code{otio_timeline}, \code{otio_track}, or \code{otio_clip}.}
}
\value{
The object's name (character).
}
\description{
Name / kind of an OTIO object
}
