% tinyrox says don't edit this manually, but it can't stop you!
\name{otio_version}
\alias{otio_version}
\title{Version of the linked OpenTimelineIO (opentime) library}
\usage{
otio_version()
}
\value{
A version string such as \code{"0.18.1"}.
}
\description{
Reports the OTIO version this build of nle.api was compiled and linked
against, useful for diagnosing which schema set is available.
}
\examples{
otio_version()
}
