% tinyrox says don't edit this manually, but it can't stop you!
\name{remove_child}
\alias{remove_child}
\title{Remove the child at a position (in place)}
\usage{
remove_child(x, index)
}
\arguments{
\item{x}{A composition or collection.}

\item{index}{1-based position.}
}
\value{
\code{x}, invisibly.
}
\description{
Remove the child at a position (in place)
}
