% tinyrox says don't edit this manually, but it can't stop you!
\name{replace_timeline_state_md}
\alias{replace_timeline_state_md}
\title{Replace only the state block in a timeline.md}
\usage{
replace_timeline_state_md(path, timeline)
}
\arguments{
\item{path}{Path to a timeline.md.}

\item{timeline}{An \code{nle_timeline}.}
}
\value{
The path, invisibly.
}
\description{
Replace only the state block in a timeline.md
}
