% tinyrox says don't edit this manually, but it can't stop you!
\docType{package}
\name{rotio-package}
\alias{rotio}
\alias{rotio-package}
\title{\packageTitle{rotio}}
\description{\packageDescription{rotio}}
\section{Package Content}{\packageIndices{rotio}}
\author{\packageAuthor{rotio}}
\section{Maintainer}{\packageMaintainer{rotio}}
\keyword{internal}
