% tinyrox says don't edit this manually, but it can't stop you!
\name{schema_version}
\alias{schema_version}
\title{Schema version of an OTIO object}
\usage{
schema_version(x)
}
\arguments{
\item{x}{An OTIO object.}
}
\value{
An integer.
}
\description{
Schema version of an OTIO object
}
\examples{
schema_version(Clip("shot1"))
}
