% tinyrox says don't edit this manually, but it can't stop you!
\name{seq_duration_frames}
\alias{seq_duration_frames}
\title{Total sequence duration in frames (end of the last clip)}
\usage{
seq_duration_frames(seq)
}
\arguments{
\item{seq}{An \code{nle_sequence}.}
}
\description{
Total sequence duration in frames (end of the last clip)
}
