% tinyrox says don't edit this manually, but it can't stop you!
\name{seq_fps}
\alias{seq_fps}
\title{Sequence frame rate as a single value (frames per second)}
\usage{
seq_fps(seq)
}
\arguments{
\item{seq}{An \code{nle_sequence}.}
}
\description{
Sequence frame rate as a single value (frames per second)
}
