% tinyrox says don't edit this manually, but it can't stop you!
\name{sequence_from_json}
\alias{sequence_from_json}
\title{Parse a cornball.sequence.v1 JSON string into an nle_sequence}
\usage{
sequence_from_json(json)
}
\arguments{
\item{json}{A JSON string.}
}
\value{
An \code{nle_sequence}.
}
\description{
Parse a cornball.sequence.v1 JSON string into an nle_sequence
}
