% tinyrox says don't edit this manually, but it can't stop you!
\name{sequence_to_json}
\alias{sequence_to_json}
\title{Serialize an nle_sequence to a cornball.sequence.v1 JSON string}
\usage{
sequence_to_json(seq, pretty = TRUE)
}
\arguments{
\item{seq}{An \code{nle_sequence}.}

\item{pretty}{Indent the output (default TRUE).}
}
\value{
A JSON string.
}
\description{
Serialize an nle_sequence to a cornball.sequence.v1 JSON string
}
