% tinyrox says don't edit this manually, but it can't stop you!
\name{set_media_references}
\alias{set_media_references}
\title{Replace the media references of a clip (in place)}
\usage{
set_media_references(x, media_references, new_active_key = NULL)
}
\arguments{
\item{x}{A \code{\link{Clip}}.}

\item{media_references}{A named list of media references.}

\item{new_active_key}{Optional new active key.}
}
\value{
\code{x}, invisibly.
}
\description{
Replace the media references of a clip (in place)
}
