% tinyrox says don't edit this manually, but it can't stop you!
\name{time_scalar}
\alias{time_scalar}
\alias{time_scalar<-}
\title{Time scalar of a LinearTimeWarp}
\usage{
time_scalar(x)

time_scalar(x) <- value
}
\arguments{
\item{x}{A \code{\link{LinearTimeWarp}}.}

\item{value}{New time scalar.}
}
\description{
Time scalar of a LinearTimeWarp
}
