% tinyrox says don't edit this manually, but it can't stop you!
\name{timeline_duration_frames}
\alias{timeline_duration_frames}
\title{Total timeline duration in frames (end of the last clip)}
\usage{
timeline_duration_frames(timeline)
}
\arguments{
\item{timeline}{An \code{nle_timeline}.}
}
\description{
Total timeline duration in frames (end of the last clip)
}
