% tinyrox says don't edit this manually, but it can't stop you!
\name{timeline_fps}
\alias{timeline_fps}
\title{Timeline frame rate as a single value (frames per second)}
\usage{
timeline_fps(timeline)
}
\arguments{
\item{timeline}{An \code{nle_timeline}.}
}
\description{
Timeline frame rate as a single value (frames per second)
}
