% tinyrox says don't edit this manually, but it can't stop you!
\name{timeline_summary}
\alias{timeline_summary}
\title{Concise human summary of a timeline}
\usage{
timeline_summary(timeline)
}
\arguments{
\item{timeline}{An \code{nle_timeline}.}
}
\description{
Concise human summary of a timeline
}
