% tinyrox says don't edit this manually, but it can't stop you!
\name{to_seconds}
\alias{to_seconds}
\title{Convert a RationalTime to seconds}
\usage{
to_seconds(x)
}
\arguments{
\item{x}{A \code{RationalTime}.}
}
\value{
Numeric scalar, the time in seconds.
}
\description{
Convert a RationalTime to seconds
}
\examples{
to_seconds(RationalTime(48, 24))
}
