% tinyrox says don't edit this manually, but it can't stop you!
\name{to_timecode}
\alias{to_timecode}
\title{SMPTE timecode for a RationalTime}
\usage{
to_timecode(x, rate = NULL, drop_frame = FALSE)
}
\arguments{
\item{x}{A \code{RationalTime}.}

\item{rate}{Timecode rate (default \code{x}'s rate).}

\item{drop_frame}{Logical; drop-frame timecode (default FALSE).}
}
\description{
SMPTE timecode for a RationalTime
}
