% tinyrox says don't edit this manually, but it can't stop you!
\name{type_version_map}
\alias{type_version_map}
\title{Current OTIO schema versions}
\usage{
type_version_map()
}
\value{
A named integer vector.
}
\description{
A named integer vector mapping each schema type to its current version,
matching OpenTimelineIO 0.18.1.
}
