% tinyrox says don't edit this manually, but it can't stop you!
\name{available_range}
\alias{available_range}
\alias{available_range<-}
\title{Available range of a media reference}
\usage{
available_range(x)

available_range(x) <- value
}
\arguments{
\item{x}{A media reference.}

\item{value}{A \code{\link{TimeRange}} or \code{NULL}.}
}
\description{
Available range of a media reference
}
