% tinyrox says don't edit this manually, but it can't stop you!
\name{clip_split}
\alias{clip_split}
\title{Split a clip at a timeline frame}
\usage{
clip_split(seq, clip, at)
}
\arguments{
\item{seq}{An \code{nle_sequence}.}

\item{clip}{Clip id.}

\item{at}{Timeline time at which to cut.}
}
\description{
Split a clip at a timeline frame
}
