% tinyrox says don't edit this manually, but it can't stop you!
\name{end_frame}
\alias{end_frame}
\title{Last frame number of an ImageSequenceReference}
\usage{
end_frame(x)
}
\arguments{
\item{x}{An \code{\link{ImageSequenceReference}}.}
}
\description{
Last frame number of an ImageSequenceReference
}
