% tinyrox says don't edit this manually, but it can't stop you!
\name{global_start_time}
\alias{global_start_time}
\alias{global_start_time<-}
\title{Global start time of a timeline}
\usage{
global_start_time(x)

global_start_time(x) <- value
}
\arguments{
\item{x}{A \code{\link{Timeline}}.}

\item{value}{A \code{\link{RationalTime}} or \code{NULL}.}
}
\description{
Global start time of a timeline
}
