% tinyrox says don't edit this manually, but it can't stop you!
\name{has_clips}
\alias{has_clips}
\title{Does an object contain any clips (recursive)?}
\usage{
has_clips(x)
}
\arguments{
\item{x}{A Timeline, composition, or collection.}
}
\description{
Does an object contain any clips (recursive)?
}
