% tinyrox says don't edit this manually, but it can't stop you!
\name{marked_range}
\alias{marked_range}
\alias{marked_range<-}
\title{Marked range of a Marker}
\usage{
marked_range(x)

marked_range(x) <- value
}
\arguments{
\item{x}{A \code{\link{Marker}}.}

\item{value}{A \code{\link{TimeRange}}.}
}
\description{
Marked range of a Marker
}
