% tinyrox says don't edit this manually, but it can't stop you!
\name{name}
\alias{name}
\alias{name<-}
\title{Get or set the name of an OTIO object}
\usage{
name(x)

name(x) <- value
}
\arguments{
\item{x}{An OTIO object.}

\item{value}{New name.}
}
\description{
Get or set the name of an OTIO object
}
