% tinyrox says don't edit this manually, but it can't stop you!
\name{tracks}
\alias{tracks}
\title{The track stack of a timeline}
\usage{
tracks(x)
}
\arguments{
\item{x}{A \code{\link{Timeline}}.}
}
\value{
The timeline's \code{\link{Stack}} of tracks.
}
\description{
The track stack of a timeline
}
