% tinyrox says don't edit this manually, but it can't stop you!
\name{transition_type}
\alias{transition_type}
\alias{transition_type<-}
\alias{in_offset}
\alias{in_offset<-}
\alias{out_offset}
\alias{out_offset<-}
\title{Transition type / offsets}
\usage{
transition_type(x)

transition_type(x) <- value

in_offset(x)

in_offset(x) <- value

out_offset(x)

out_offset(x) <- value
}
\arguments{
\item{x}{A \code{\link{Transition}}.}

\item{value}{New value.}
}
\description{
Transition type / offsets
}
