% tinyrox says don't edit this manually, but it can't stop you!
\name{trimmed_range_in_parent}
\alias{trimmed_range_in_parent}
\title{Range of an item within its parent, trimmed by the parent's source range}
\usage{
trimmed_range_in_parent(x)
}
\arguments{
\item{x}{An item with a parent.}
}
\value{
A \code{\link{TimeRange}}.
}
\description{
Range of an item within its parent, trimmed by the parent's source range
}
